--- events/guildMemberAdd.js
@@ -1,6 +1,0 @@
-module.exports = async (client, user) => {
-	if (client.userDBArchive.has(user.id)) {
-		client.userDB.set(user.id, client.userDB.get(user.id));
-		client.userDBArchive.delete(user.id);
-	}
-};
